backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.55.0
python-dotenv>=1.0.0
numpy>=1.26.0
faiss-cpu>=1.12.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
